feature(Transformations): added `AddRollingCorrelation` class
