--- src/fold/transformations/window.py
@@ -1,121 +1,0 @@
-# Copyright (c) 2022 - Present Myalo UG (haftungbeschränkt) (Mark Aron Szulyovszky, Daniel Szemerey) <info@dreamfaster.ai>. All rights reserved. See LICENSE in root folder.
-
-
-from __future__ import annotations
-
-from typing import Callable, List, Optional, Tuple, Union
-
-import pandas as pd
-
-from ..base import Artifact, Transformation, Tunable, fit_noop
-from ..utils.enums import ParsableEnum
-from ..utils.list import wrap_in_list
-
-
-class PredefinedFunction(ParsableEnum):
-    mean = "mean"
-    sum = "sum"
-    median = "median"
-    std = "std"
-    var = "var"
-    kurt = "kurt"
-    min = "min"
-    max = "max"
-    corr = "corr"
-    cov = "cov"
-    skew = "skew"
-    sem = "sem"
-
-
-ColumnOrColumns = Union[str, List[str]]
-FunctionOrPredefined = Union[Callable, PredefinedFunction, str]
-ColumnWindowFunction = Tuple[ColumnOrColumns, int, FunctionOrPredefined]
-
-
-class AddWindowFeatures(Transformation, Tunable):
-
-    """
-    Creates rolling window features on the specified columns.
-    Equivalent to adding a new column by running: `df[column].rolling(window).function()`.
-
-
-    Parameters
-    ----------
-    column_window_func : ColumnWindowFunction, List[ColumnWindowFunction]
-        A list of tuples, where each tuple contains the column name, the window size and the function to apply.
-        The function can be a predefined function (see PredefinedFunction) or a Callable (with a single parameter).
-
-    Examples
-    --------
-    ```pycon
-    >>> from fold.loop import train_backtest
-    >>> from fold.splitters import SlidingWindowSplitter
-    >>> from fold.transformations import AddWindowFeatures
-    >>> from fold.utils.tests import generate_sine_wave_data
-    >>> X, y  = generate_sine_wave_data()
-    >>> splitter = SlidingWindowSplitter(train_window=0.5, step=0.2)
-    >>> pipeline = AddWindowFeatures(("sine", 10, "mean"))
-    >>> preds, trained_pipeline = train_backtest(pipeline, X, y, splitter)
-    >>> preds.head()
-                           sine  sine_10_mean
-    2021-12-31 15:40:00 -0.0000      -0.05649
-    2021-12-31 15:41:00  0.0126      -0.04394
-    2021-12-31 15:42:00  0.0251      -0.03139
-    2021-12-31 15:43:00  0.0377      -0.01883
-    2021-12-31 15:44:00  0.0502      -0.00628
-
-    ```
-    """
-
-    def __init__(
-        self,
-        column_window_func: Union[ColumnWindowFunction, List[ColumnWindowFunction]],
-        name: Optional[str] = None,
-        params_to_try: Optional[dict] = None,
-    ) -> None:
-        self.column_window_func = [
-            (
-                wrap_in_list(column),
-                window,
-                function
-                if isinstance(function, Callable)
-                else PredefinedFunction.from_str(function),
-            )
-            for column, window, function in wrap_in_list(column_window_func)
-        ]
-        max_memory = max([window for _, window, _ in self.column_window_func])
-        self.properties = Transformation.Properties(
-            requires_X=True, memory_size=max_memory
-        )
-        self.params_to_try = params_to_try
-        self.name = name or "AddWindowFeatures"
-
-    def transform(
-        self, X: pd.DataFrame, in_sample: bool
-    ) -> Tuple[pd.DataFrame, Optional[Artifact]]:
-        X_function_applied = pd.DataFrame([], index=X.index)
-        for columns, window, function in self.column_window_func:
-            if isinstance(function, PredefinedFunction):
-                function = getattr(pd.core.window.rolling.Rolling, function.value)
-            function_name = (
-                function.__name__ if function.__name__ != "<lambda>" else "transformed"
-            )
-
-            if columns[0] == "all":
-                columns = X.columns
-
-            X_function_applied = pd.concat(
-                [
-                    X_function_applied,
-                    function(
-                        X[columns]
-                        .add_suffix(f"_{window}_{function_name}")
-                        .rolling(window)
-                    ),
-                ],
-                axis="columns",
-            )
-        return pd.concat([X, X_function_applied], axis="columns"), None
-
-    fit = fit_noop
-    update = fit
